refactor: mudulfy Menu
